feat(core): rename the json fonction to jsonDecode
